+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 Change Log
 ==========
+## 3.2.0
+- add `jpex.raw` function for extracting a factory function
+
 ## 3.1.0
-- global types like `Window` and `Document` and now be used to register dependencies
+- global types like `Window` and `Document` can now be used to register dependencies
 
 ## 3.0.1
 - encase now caches the wrapped function for better performance